+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Changed Configuration layout to be more readable, by using headers and colours to separate sections.
+- Changed "Golden Slime Rain" to increase spawn-rates if the "Increase spawn-rate" effect is active.
 - Fixed "On Fire" de-buff not bypassing immunities.
 
 ## [1.0.3] - 2022-07-29